safari-600.7-branch/LayoutTests/inspector-obsolete/debugger/debugger-scripts-reload-expected.txt

Tests that scripts list is cleared upon page reload.

Debugger was enabled.
Dummy script found: dummyScript.js
Page reloaded.
Debugger was disabled.
